handle enterance of one user with different devices

+++ src/main/java/com/rahnemacollege/configuration/WebSocketConfig.java
@@ -6,14 +6,17 @@ import com.rahnemacollege.service.UserService;
 import com.rahnemacollege.util.JwtTokenUtil;
 import com.rahnemacollege.util.exceptions.MessageException;
 import org.slf4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.messaging.simp.stomp.StompCommand;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.messaging.support.ChannelInterceptor;
+import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.messaging.support.MessageHeaderAccessor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -33,6 +36,9 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     private UserService userService;
     private BidService bidService;
 
+    @Autowired
+    private SimpMessagingTemplate template;
+
     public WebSocketConfig(JwtTokenUtil tokenUtil, UserService userService, BidService bidService) {
         this.tokenUtil = tokenUtil;
         this.userService = userService;
@@ -45,7 +51,7 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     @Override
     public void configureMessageBroker(MessageBrokerRegistry registry) {
         registry.setApplicationDestinationPrefixes("/auction", "/app") //socket_subscriber
-                .enableSimpleBroker("/app"); //socket_publisher
+                .enableSimpleBroker("/app", "/auction"); //socket_publisher
     }
 
     @Override
@@ -64,19 +70,26 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
                 if (headerAccessor != null && StompCommand.CONNECT.equals(headerAccessor.getCommand())) {
                     logger.info("try to connect");
                     String jwtToken = Objects.requireNonNull(headerAccessor.getFirstNativeHeader("auth")).substring(7);
+                    String deviceId = Objects.requireNonNull(headerAccessor.getFirstNativeHeader("deviceID"));
                     String id = tokenUtil.getIdFromToken(jwtToken);
                     if (tokenUtil.isTokenExpired(jwtToken))
                         throw new MessageException(com.rahnemacollege.util.exceptions.Message.TOKEN_NOT_FOUND);
                     User user = userService.findUserId(Integer.valueOf(id));
-                    Authentication u = new UsernamePasswordAuthenticationToken(user.getId().toString(), user.getPassword(), new ArrayList<>());
+                    bidService.addDeviceId(deviceId, user);
+                    Authentication u = new UsernamePasswordAuthenticationToken(deviceId, user.getPassword(), new ArrayList<>());
                     headerAccessor.setUser(u);
                     logger.info("the user with session id " + headerAccessor.getSessionId() + " connected");
-                } else if (headerAccessor != null && StompCommand.SUBSCRIBE.equals(headerAccessor.getCommand())) {
+                }
+                if (headerAccessor != null && StompCommand.SUBSCRIBE.equals(headerAccessor.getCommand())) {
                     System.err.println(headerAccessor.getDestination());
-                } else if (headerAccessor != null && StompCommand.DISCONNECT.equals(headerAccessor.getCommand()) && headerAccessor.getUser() != null) {
-
-                    User user = userService.findUserId(Integer.valueOf(headerAccessor.getUser().getName()));
+                }
+                if (headerAccessor != null && StompCommand.DISCONNECT.equals(headerAccessor.getCommand()) && headerAccessor.getUser() != null) {
+                    Integer userId = bidService.getUserId(headerAccessor.getUser().getName());
+                    if (userId == null)
+                        return message;
+                    User user = userService.findUserId(userId);
                     bidService.removeFromAllAuction(user);
+                    bidService.removeDeviceId(headerAccessor.getUser().getName());
                 }
                 return message;
             }
@@ -84,5 +97,4 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     }
 
 }
-
