+++ src/main/java/com/rahnemacollege/controller/SessionUnsubscribeListener.java
@@ -41,7 +41,12 @@ public class SessionUnsubscribeListener implements ApplicationListener<SessionUn
         logger.info("someone try to exit from auction");
         GenericMessage message = (GenericMessage) event.getMessage();
         StompHeaderAccessor headerAccessor = MessageHeaderAccessor.getAccessor(message, StompHeaderAccessor.class);
-        User user = userService.findUserId(Integer.parseInt(headerAccessor.getUser().getName()));
+        Integer userId = bidService.getUserId(headerAccessor.getUser().getName());
+        if(userId == null){
+            logger.error("the user is null");
+            return;
+        }
+        User user = userService.findUserId(userId);
         String subscriptionId = String.valueOf(user.getId());
         Subscription subscription = bidService.getSubscription(subscriptionId);
         bidService.removeFromAllAuction(user);
